requirements.txt
asgiref==3.5.2
auxlib==0.0.43
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
distlib==0.3.6
Django==4.1.1
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
filelock==3.8.0
idna==3.4
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
MarkupSafe==2.1.1
packaging==21.3
platformdirs==2.5.2
psycopg2-binary==2.9.3
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.1
